style: clean up unused codes
