+++ bukkit-api/src/main/kotlin/chuyong/springspigot/child/SpringSpigotPluginRegistry.kt
