--- paper-api/src/main/resources/plugin.yml
@@ -1,17 +1,0 @@
-name: SpigotSpring
-author: CChuYong
-website: https://chuyong.kr
-version: '${version}'
-main: chuyong.springspigot.SpringSpigotBootstrapper
-api-version: 1.17
-load: STARTUP
-depend:
-  - HQFramework
-libraries:
-  - org.apache.commons:commons-lang3:3.12.0
-  - com.github.f4b6a3:ulid-creator:5.2.0
-  - mysql:mysql-connector-java:8.0.28
-  - com.h2database:h2:1.4.200
-  - net.bytebuddy:byte-buddy:1.14.5
-  - org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.7.21
-  - org.jetbrains.kotlin:kotlin-reflect:1.8.0
