+++ settings.gradle.kts
@@ -1,3 +1,3 @@
 rootProject.name = "spigot-spring"
-include("core-api", "bukkit-class-modifier", "spigot-api", "bukkit-api", "paper-api")
+include("core-api", "bukkit-class-modifier", "bukkit-api")
 include("bukkit-bootstrapper")
