--- spigot-api/src/main/kotlin/chuyong/springspigot/EnableSpringSpigotSupport.kt
@@ -1,7 +1,0 @@
-package chuyong.springspigot
-
-@Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
-annotation class EnableSpringSpigotSupport(
-    val contextEscalation: Boolean = false,
-)
