--- spigot-api/src/main/kotlin/chuyong/springspigot/MockMain.kt
@@ -1,5 +1,0 @@
-package chuyong.springspigot
-
-import org.bukkit.plugin.java.JavaPlugin
-
-class MockMain : JavaPlugin()
