--- spigot-api/src/main/kotlin/chuyong/springspigot/SpringSpigotApplication.kt
@@ -1,11 +1,0 @@
-package chuyong.springspigot
-
-import org.springframework.context.annotation.ComponentScan
-import org.springframework.context.annotation.Configuration
-import org.springframework.scheduling.annotation.EnableScheduling
-
-
-@EnableScheduling
-@ComponentScan(basePackages = ["chuyong.springspigot"])
-@Configuration
-class SpringSpigotApplication
